requirements.txt
click==8.1.1
datetime==4.3
flask_cors==3.0.10
Flask==2.0.3
itsdangerous==2.1.2
Jinja2==3.1.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
MarkupSafe==2.1.1
markovify==0.9.4
python-dotenv==0.21.0
re==2.2.1
requests==2.25.1
Werkzeug==2.1.0
